+++ packages/harpstrata/src/harpstrata/Covariant/deduceHarpKeyId/index.ts
@@ -1,12 +1,12 @@
 import { HarpKeyControllers } from '../types'
-import { getPozitionRootOffset } from '../../Pozition'
+import { getPozition } from '../../Pozition'
 import type { PitchIds } from '../../Pitch'
 import { getDescendingPitchIds } from '../../OrderedIds'
 
 
 export const deduceHarpKeyId = (props: HarpKeyControllers): PitchIds => {
   const { rootPitchId, pozitionId } = props
-  const rootOffset = getPozitionRootOffset(pozitionId)
+  const { rootOffset } = getPozition(pozitionId)
 
   const descendingPitchIds = getDescendingPitchIds(rootPitchId)