+++ packages/harpstrata/src/harpstrata/Covariant/deducePozitionId/index.ts
@@ -1,6 +1,6 @@
 import { PozitionControllers } from '../types'
+import { getPozitionByOffset } from '../../Pozition'
 import type { PozitionIds } from '../../Pozition'
-import { getAscendingPozitionIds } from '../../OrderedIds'
 import { getAscendingPitchIds } from '../../OrderedIds'
 
 
@@ -9,9 +9,9 @@ export const deducePozitionId = (props: PozitionControllers): PozitionIds => {
 
   const harpKeyAscendingPitchIds = getAscendingPitchIds(harpKeyId)
 
-  const rootPitchIndex = harpKeyAscendingPitchIds.indexOf(rootPitchId)
+  const rootPitchOffset = harpKeyAscendingPitchIds.indexOf(rootPitchId)
 
-  const { [rootPitchIndex]: pozitionId } = getAscendingPozitionIds()
+  const { id: pozitionId } = getPozitionByOffset(rootPitchOffset)
 
   return pozitionId
 }