+++ packages/harpstrata/src/harpstrata/Covariant/deduceRootPitchId/index.ts
@@ -1,12 +1,12 @@
 import { RootPitchControllers } from '../types'
-import { getPozitionRootOffset } from '../../Pozition'
+import { getPozition } from '../../Pozition'
 import type { PitchIds } from '../../Pitch'
 import { getAscendingPitchIds } from '../../OrderedIds'
 
 
 export const deduceRootPitchId = (props: RootPitchControllers): PitchIds => {
   const { harpKeyId, pozitionId } = props
-  const rootOffset = getPozitionRootOffset(pozitionId)
+  const { rootOffset } = getPozition(pozitionId)
 
   const ascendingPitchIds = getAscendingPitchIds(harpKeyId)