+++ packages/harpstrata/src/harpstrata/HarpStrata/getHarpStrata/index.ts
@@ -1,5 +1,5 @@
 import type { HarpStrataProps, HarpStrata } from '../types'
-import { getPozitionRootOffset } from '../../Pozition'
+import { getPozition } from '../../Pozition'
 import { getPitchMatrix, getPitch } from '../../Pitch'
 import { getIsActiveComplex } from '../../IsActive'
 import { getDegreeMatrix } from '../../Degree'
@@ -10,13 +10,13 @@ export const getHarpStrata = (props: HarpStrataProps): HarpStrata => {
   const { apparatusId, pozitionId, harpKeyId, activeIds } = props
 
   const apparatus = getApparatus(apparatusId)
-  const pozitionRoot = getPozitionRootOffset(pozitionId)
+  const { rootOffset } = getPozition(pozitionId)
   const pitch = getPitch(harpKeyId)
 
   const { halfstepIndexMatrix } = apparatus
   const { id: pitchId } = pitch
 
-  const degreeMatrix = getDegreeMatrix(halfstepIndexMatrix, pozitionRoot)
+  const degreeMatrix = getDegreeMatrix(halfstepIndexMatrix, rootOffset)
   const pitchMatrix = getPitchMatrix(halfstepIndexMatrix, pitchId)
   const isActiveComplex = getIsActiveComplex({degreeMatrix, pitchMatrix, activeIds: activeIds })