+++ packages/harpstrata/src/harpstrata/Pozition/getPozitionRootOffset/index.ts
@@ -1,4 +1,0 @@
-import { PozitionIds } from '../../Pozition'
-import { getAscendingPozitionIds } from '../../OrderedIds'
-
-export const getPozitionRootOffset = (pozitionId: PozitionIds): number => getAscendingPozitionIds().indexOf(pozitionId)