+++ packages/harpstrata/src/harpstrata/Pozition/index.ts
@@ -1,5 +1,4 @@
-export { getPozition } from './getPozition'
-export { getPozitionRootOffset } from './getPozitionRootOffset'
+export { getPozition, getPozitionByOffset } from './getPozition'
 
 export { PozitionIds } from './types'
 export type { Pozition } from './types'